moved tests and deleted unnecessary ones
