+++ test-app/assets/app/tests/testArrays.js
@@ -7,7 +7,32 @@ describe("Tests array operations", function () {
 	beforeEach(function() {
 		jasmine.addCustomEqualityTester(myCustomEquality);
 	});
-	
+
+	it("TestWorkingWithJavaArrayDoesNotMakeMemoryLeak", function () {
+
+		__log("TEST: TestWorkingWithJavaArrayDoesNotMakeMemoryLeak");
+		
+		var size = 10 * 1024 * 1024;
+		
+		for (var i = 0; i < 100; i++) {
+		
+			var arr = java.lang.reflect.Array.newInstance(java.lang.Byte.class.getField("TYPE").get(null), size);
+			
+			var length = arr.length;
+			
+			expect(length).toEqual(size);
+			
+			arr[0] = 123;
+			
+			var el = arr[0];
+			
+			expect(el).toEqual(123);
+		
+			gc();
+			java.lang.System.gc();
+		}
+	});
+
 	it("TestArraySize", function () {
 		
 		var size = 12345;
@@ -36,4 +61,61 @@ describe("Tests array operations", function () {
 		}
 	});
 
+	it("TestArrays", function () {
+		
+		__log("TEST: TestArrays");
+		
+		var MyButton = com.tns.tests.Button1.extend("MyButton639", {
+			toString : function() {
+			  	return "button1";	
+			}
+		});
+		var tester = new MyButton(); 
+		var instances = tester.getDummyInstances();
+		
+		var instanceFound = false;
+
+		for (var i = 0; i < instances.length; i++)
+		{
+			if (instances[i].getName() == "second");
+			{
+				instanceFound = true;
+			}
+		}
+		
+		expect(instanceFound).toEqual(true);
+		
+		instances[0] = instances[1];
+		
+		var instances0name = instances[0].getName();
+		var instances1name = instances[1].getName();
+		
+		expect(instances0name).toEqual(instances1name);
+	});
+	
+	it("TestArrayLengthPropertyIsNumber", function () {
+		
+		__log("TEST: TestArrayLengthPropertyIsNumber");
+		
+		var expectedLength = 10;
+
+		function getLength(x)
+		{
+			var arr = x.getIntArray1(expectedLength);
+			
+			return arr ? arr.length : 123456;
+		}
+		
+		var MyButton = com.tns.tests.Button1.extend("MyButton680", {
+			toString : function() {
+			  	return "button1";	
+			}
+		});
+		
+		var count = getLength(new MyButton());
+		
+		expect(count).toBe(expectedLength);
+		
+	});
+	
 });
